Some adjustments to new overflow altars
For nzn_hepliaklqana_mist_gate and nzn_wu_jian_contemplation some minor
simplifications and spacing cleanups.

For nzn_makhleb_speed_demon, replace the executioner with a sixfirhy,
since an executioner is basically worse than any possible 8 spawn for
D:7-D:11. A sixfirhy is quite fast and dangerous for these same depths,
but not so much that no character would reasonably open the vault to get
to the Makhleb altar.

For nzn_overflow_gifts_of_might_and_magic, revamp the Trog-themed weapon
to add enchant at the earliest depth and have Trog-themed brands (so
vorpal, flaming, or antimagic) but no enchant at subsequent depths. This
way the Trog item offering isn't so meagre compared to the Sif item,
which gifts a small number of low-level spells.