Put the cursor back in the X menu on Windows (#1536)
See the comments on the PR. As aidanh mentions, cursor_control is a
blunt force instrument that needs to be manually turned on and off here,
but my naive attempts to enable the cursor only when the map cursor is
being drawn didn't work -- it seems that it gets redrawn at odd moments,
and that ends up wiping the cursor away.

I didn't bother to make use_fake_cursor work here, because it already
was broken for the direction chooser anyway. If we ever hear about
someone using it on Windows, we can fix it, but it seems like it only
fixes a Unix-specific problem.

I'd wait for a more thorough fix, but this is a game-breaking bug on
Windows console and we have a release coming up.

Fixes #1536

This also incidentally fixes a related problem where the cursor appeared
when opening up sub-menus from the direction chooser.