Fix a butterfly+charge+quiver interaction (12410)
Palentonga charge can target butterflies, but autofight code has its own
check that ignores butterflies. With autofight_throw=true, when a
butterfly was in charge range the quivered ability would therefore be
enabled, but autofight would select movement towards the distant target.
However, autofight relies on a quivered ability to be disabled to know
when to trigger movement; the net result was that it passed a distant
target through to activate the ability and gave an error.

On the assumption that players won't usually want to roll towards a
butterfly, this commit disables charge with non-threatening monsters in
range. However, charge can still be triggered towards a non-threatening
monster by using the `a` menu. As noted in the comments this is a
somewhat hacky solution, and a more general solution along the lines of
spells + non-threatening monsters would be less hacky.