Fix launcher autoswitch issues (ardl)
launcher_action_cycler autoswitch was broken in some cases (most
obviously when picking up a launcher for the first time with none of its
ammo in inventory) because its default initialization was with an
ammo_action, not a launcher_ammo_action.  While debugging this I found a
bunch of related issues and cleaned up some of the code slightly.