Fixup for unsaved abyss crashes
If the game crashes without saving before an AK start has saved and
while they are in the abyss, the abyss map is not saved, and they reload
in an empty featureless void (that crashes). This fixup code simply adds
a floor and moves the player to that position, as a very ad-hoc
emergency measure so that the game will load. The results look weird but
should allow the AK start to get out of the abyss (the position is also
one that triggers an abyss shift immediately if they want to stick
around).