Further work to make the distortion unwield rework function properly on gloves of wielding. Change distortion unwield effects to: 1/3 Nothing. 1/6 3-12 Damage + Blink (mostly harmless) 1/6 Banish 1/6 Summon Spatial Vortices 1/6 Contam (Not as random as before, always Yellow; from 5000 (minimum Yellow) to 10000 (Halfway between Yellow and Lightred).
Removed the blink to monsters effect as that one was a bit of a spoilery "go to a floor with no monsters to unwield" or even a
useful "go to a floor with just a unique left and use an unwield to find it"

Fairly harmless effects from low severity miscasts readded. Spatial Vortices effect from severe miscast readded. Contam made a bit less
true random (0 to Red levels at random, really?).